+++ quadEMApp/src/four1.c
@@ -37,8 +37,8 @@ void four1(float data[], unsigned long nn, int isign)
 		for (m=1;m<mmax;m+=2) {
 			for (i=m;i<=n;i+=istep) {
 				j=i+mmax;
-				tempr=wr*data[j]-wi*data[j+1];
-				tempi=wr*data[j+1]+wi*data[j];
+				tempr=(float)(wr*data[j]-wi*data[j+1]);
+				tempi=(float)(wr*data[j+1]+wi*data[j]);
 				data[j]=data[i]-tempr;
 				data[j+1]=data[i+1]-tempi;
 				data[i] += tempr;
